Декомпозиция калькулятора по mvc
