+++ src/ui/main_window.py
@@ -1,11 +1,11 @@
 from PyQt6.QtWidgets import QTabWidget, QVBoxLayout, QWidget
 
+from src.controllers.calculator_controller import CalculatorController
 from src.controllers.package_controller import PackageController
 from src.controllers.settings_controller import SettingsController
+from src.ui.tabs.calculator_tab import CalculatorTab
 from src.ui.tabs.package_tab import PackageTab
 from src.ui.tabs.settings_tab import SettingsTab
-
-from src.calculator.main import CalculatorWindow
 
 
 class DocPrepApp(QWidget):
@@ -13,26 +13,26 @@ class DocPrepApp(QWidget):
         super().__init__()
         self.setWindowTitle('DocPrep')
         self.resize(600, 600)
-        self.setting_controller = None
         self.package_controller = None
+        self.setting_controller = None
+        self.calculator_controller = None
         self.init_ui()
 
     def init_ui(self):
         layout = QVBoxLayout()
         tabs = QTabWidget()
 
+        calculator_tab = CalculatorTab()
         package_tab = PackageTab()
         settings_tab = SettingsTab()
-        
-        calculator_tab = CalculatorWindow()
 
         # Контроллеры
         self.package_controller = PackageController(package_tab)
         self.setting_controller = SettingsController(settings_tab)
+        self.calculator_controller = CalculatorController(calculator_tab)
 
         tabs.addTab(package_tab, 'Пакет документов')
         tabs.addTab(settings_tab, 'Настройки')
-        
         tabs.addTab(calculator_tab, 'Калькулятор')
 
         layout.addWidget(tabs)
